add jquery-validation

+++ SoundPlay/SoundPlay.DAL/Models/Guitar.cs
@@ -4,8 +4,8 @@ namespace SoundPlay.DAL.Models
 {
     public sealed class Guitar : Product
     {
-        public int FretsCount { get; set; }
-        public int StringsCount { get; set; }
+        public byte FretsCount { get; set; }
+        public byte StringsCount { get; set; }
         public int? ShapeId { get; set; }
         public int? SoundboardId { get; set; }
         public int? NeckId { get; set; }
